rock/requirements.txt
Django==2.2.2
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-formtools==2.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
django-extensions==2.2.1
django-filter==2.0.0
django-filters==0.2.1
mysqlclient==1.4.2.post1
PyJWT==1.7.1
PyMySQL==0.7.11
requests==2.22.0
SQLAlchemy==1.1.5
sqlparse==0.3.0
xlrd==1.0.0
XlsxWriter==0.9.6
xlwt==1.2.0
django-werkzeug-debugger-runserver==0.3.1
pyOpenSSL==16.2.0
Pillow==5.4.1

## wsgi
uWSGI==2.0.17.1
uwsgidecorators==1.1.0